+++ _import/tycho-execution/src/encoding/evm/swap_encoder/mod.rs
@@ -9,6 +9,7 @@ mod etherfi;
 mod fluid_v1;
 mod hashflow;
 mod liquidity_party;
+mod liquorice;
 mod maverick_v2;
 mod rocketpool;
 mod slipstreams;