+++ _import/tycho-execution/src/encoding/evm/swap_encoder/swap_encoder_registry.rs
@@ -11,7 +11,8 @@ use crate::encoding::{
             bebop::BebopSwapEncoder, curve::CurveSwapEncoder, ekubo::EkuboSwapEncoder,
             ekubo_v3::EkuboV3SwapEncoder, erc_4626::ERC4626SwapEncoder,
             etherfi::EtherfiSwapEncoder, fluid_v1::FluidV1SwapEncoder,
-            hashflow::HashflowSwapEncoder, liquidity_party::LiquidityPartySwapEncoder,
+            liquidity_party::LiquidityPartySwapEncoder,
+            hashflow::HashflowSwapEncoder, liquorice::LiquoriceSwapEncoder,
             maverick_v2::MaverickV2SwapEncoder, rocketpool::RocketpoolSwapEncoder,
             slipstreams::SlipstreamsSwapEncoder, uniswap_v2::UniswapV2SwapEncoder,
             uniswap_v3::UniswapV3SwapEncoder, uniswap_v4::UniswapV4SwapEncoder,
@@ -136,6 +137,9 @@ impl SwapEncoderRegistry {
             }
             "rfq:hashflow" => {
                 Ok(Box::new(HashflowSwapEncoder::new(executor_address, self.chain, config)?))
+            }
+            "rfq:liquorice" => {
+                Ok(Box::new(LiquoriceSwapEncoder::new(executor_address, self.chain, config)?))
             }
             "fluid_v1" => {
                 Ok(Box::new(FluidV1SwapEncoder::new(executor_address, self.chain, config)?))